add support for websocket ping (#1)
A compliant OCPP client has to support the WebSocketPingInterval configuration
and thus, has to be able to send websocket ping at regular interval.

This adds a send_ping() method allowing a client to send a websocket ping
and to wait for the pong reply.

It also adds an on_ping() method allowing the client to be notified when
receiving a websocket ping from the server. The websocket pong reply is sent
automatically by tungstenite.

Co-authored-by: Aurélien Jacobs <aj@techno-city.fr>